refactor: camelCase util files
