+++ src/globals/fs-path.d.ts
@@ -1,3 +1,3 @@
-import { fsPath as _fsPath } from '@/utils/fs-path';
+import { fsPath as _fsPath } from '@/utils/fsPath';
 export as namespace fsPath;
 export = _fsPath;
