+++ src/globals/jsx-to-html.d.ts
@@ -1,3 +1,3 @@
-import { jsxToHTML as _jsxToHTML } from '@/utils/jsx-to-html';
+import { jsxToHTML as _jsxToHTML } from '@/utils/jsxToHTML';
 export as namespace jsxToHTML;
 export = _jsxToHTML;
