+++ src/utils/cachebuster.ts
@@ -1,4 +1,4 @@
-import { emptyObject } from '@/utils/empty-object';
+import { emptyObject } from '@/utils/emptyObject';
 
 export function cacheBuster(_filepath: string): boolean {
 	const filepath = _filepath.startsWith('/') ? _filepath : require.resolve(_filepath);
