+++ src/utils/emptyObject.ts
