+++ src/utils/logger.ts
@@ -1,7 +1,7 @@
 /* eslint-disable no-console */
 import fsSync from 'fs';
 import { inspect } from 'util';
-import { fsPath } from '@/utils/fs-path';
+import { fsPath } from '@/utils/fsPath';
 
 function dimText(str: string): string {
 	return `\x1b[2m${str}\x1b[22m`;
